+++ e2e-tests/tests/api.rs
@@ -1,4 +1,5 @@
 use candid::Principal;
+use ic_cdk::management_canister::{CanisterSettings, EnvironmentVariable, UpdateSettingsArgs};
 use pocket_ic::ErrorCode;
 
 mod test_utilities;
@@ -144,6 +145,7 @@ fn call_api() {
         .update_call(canister_id, sender, "call_cost_sign_with_schnorr", vec![])
         .unwrap();
     assert!(res.is_empty());
+
     let res = pic
         .update_call(canister_id, sender, "call_debug_print", vec![])
         .unwrap();
@@ -153,4 +155,57 @@ fn call_api() {
         .unwrap_err();
     assert_eq!(rej.error_code, ErrorCode::CanisterCalledTrap);
     assert!(rej.reject_message.contains("It's a trap!"));
+}
+
+// Test for the system API that are not enabled on the mainnet yet.
+// A canister Wasm module that imports the non-mainnet APIs would result in canister installation failure.
+#[test]
+fn call_api_nonmainnet() {
+    let wasm = cargo_build_canister("api_nonmainnet");
+    let pic = pic_base().with_nonmainnet_features(true).build();
+    let canister_id = pic.create_canister();
+    pic.add_cycles(canister_id, 100_000_000_000_000);
+    pic.install_canister(canister_id, wasm, vec![], None);
+    let sender = Principal::anonymous();
+    let update_settings_arg = UpdateSettingsArgs {
+        canister_id,
+        settings: CanisterSettings {
+            environment_variables: Some(vec![
+                EnvironmentVariable {
+                    name: "key1".to_string(),
+                    value: "value1".to_string(),
+                },
+                EnvironmentVariable {
+                    name: "key2".to_string(),
+                    value: "value2".to_string(),
+                },
+            ]),
+            ..Default::default()
+        },
+    };
+    let _: () = update(
+        &pic,
+        Principal::management_canister(),
+        "update_settings",
+        (update_settings_arg,),
+    )
+    .unwrap();
+
+    // As of 2025-08, the env_var APIs are not available on mainnet yet.
+    let res = pic
+        .update_call(canister_id, sender, "call_env_var_count", vec![])
+        .unwrap();
+    assert!(res.is_empty());
+    let res = pic
+        .update_call(canister_id, sender, "call_env_var_name", vec![])
+        .unwrap();
+    assert!(res.is_empty());
+    let res = pic
+        .update_call(canister_id, sender, "call_env_var_name_exists", vec![])
+        .unwrap();
+    assert!(res.is_empty());
+    let res = pic
+        .update_call(canister_id, sender, "call_env_var_value", vec![])
+        .unwrap();
+    assert!(res.is_empty());
 }