+++ library/ic-certified-map/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Implement CandidType, Serialize, and Deserialize for the RbTree.
+- Implement Deserialize for the HashTree.
 
 ## [0.4.0] - 2023-07-13